% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChileAirQualityApp.R
\name{ChileAirQualityApp}
\alias{ChileAirQualityApp}
\title{Title ChileAirQualityApp}
\usage{
ChileAirQualityApp()
}
\value{
A shiny dashboard to work with the package
}
\description{
This tool is a dashboard that allows you to use the data download
 functions of this package enhanced with analysis, visualization and
 descriptive statistics tools.
}
\examples{
\dontrun{ChileAirQualityApp()}
}
\seealso{
<https://chileairquality.shinyapps.io/chileairquality/>
}
